updated with new UI and includes university query
Tutorial was updated with the information from the original user profile tutorial for query including using the pre-build database universities.  The new UI was created to handle searching for the universities that uses the storyboard file and constraints to give a similar user experience with less code and more standard code.